Add method to get output properties
Add method that makes it possible to get x-offset, y-offset, width,
height and some more properties of any specific output (by name).
This uses the root window, so it only needs connection and status to get
the info for the output. If the output cannot be found, it will return
an error.