Fix window properties
Due to incorrect type atoms the window properties have been set
incorrectly. To fix this the `set_ewmh_prop!` macro has been reworked.
It now is a lot more flexible, which also lead to the `set_prop!`
rename. This fixes #37.

Because this now allows easier change of properties I've also set the
window title property using both `_NET_WM_NAME` and `WM_NAME` so this
should be quite the simple fix for #33 too.