Improve Travis CI integration (#40)
* Add basic Time component

Added a basic time component that doesn't rely on the `date` command.
This is an initial prototype for the smart clock described in
dogamak/xcbars#3. It chooses the refresh rate of one minute/one second
depending on the time format.

* Add clippy to travis configuration

Travis should now check all clippy lints and deny the build if there has
been a problem with clippy.

* Add Nightly Docker file

Added two separate configurations for Nightly and Stable docker
containers. This makes it possible to test the most important build
(Stable), but still allows checking Clippy and the Nightly build.

* Fix Docker tags

Docker tags need to be lowercase. This has been fixed in the travis
config.

* Change docker configuration

My shady online sources are telling me I missed a `.` in the docker
command. This should get the build going now with two docker containers.

* Add rustfmt to travis configuration

Travis configuration now also checks if rustfmt exits with a 0 exit
code. This makes sure that there is no need to format a lot of code
after it has already been merged.

* Test Travis build matrix

Setup a basic configuration using the matrix feature.
Hopefully this will lead to multiple builds being triggered instead of
just one.

* Specify $COMMAND as String

The env variables only contain the first word of multiple words
separated with spaces. Now with the whole thing as one String this might
be fixed.

* Fix env specification

Now the env variables are properly specified. This will prevent stable
from building with rustfmt or clippy.

* Add cargo bin directory to path

Rustfmt is complaining about not being inside of $PATH. This adds
`$HOME/.cargo/bin` to the `$PATH` for the nightly build.

* Remove rustfmt

Rustfmt has been creating some issues, complaining about missing PATH.
To make sure the multiple builds can work this has been removed.

* Add DOCKERFILE to env

Modified the travis configuration again so it pulls only sets up the
required dockerfile. This might also fix the problem with only a single
build running.

* Fix Travis before_install configuration

The before_install was still using a env variable that has been deleted.
That variable has been replaced with a fixed value.

* Use remote docker registry

To decrease chances of failing Debian mirrors the docker file is now
pulled from the remote `undeadleech/xcbars` registry.
This should hopefully make the container fail less often during package
install.

* Inline install_rust.sh commands

Docker seems to have trouble reading commands from files. So for testing
purposes all commands have been inlined.

The language has also changed from `Rust` to `generic` to prevent
installing unnecessary applications.

* Extract Rust install to file

Travis doesn't like a lot of inline commands for docker. So the whole
`docker run …` block has been extracted to a file.

* Add `sh` command to install script

* Fix docker image name

The image name in `script` was still the old `dogamak/xcbars`. Using the
new `undeadleech/xcbars` will correctly use the image that has been
pulled in `before_install`.

* Commit changes made in the `before_install` script

To install Rust there needs to be a change made to the actual image, not
just a temporary change in the image. To do this the changes made in the
`install_rust.sh` script are now applied to the docker image using the
`docker commit …` command.

* Add clippy and fmt to nightly

Added a second run/commit step that adds clippy and rustfmt-nightly to
the image if the `RUSTV` variable is `nightly`.

* Switch before_install script to bash

sh had problems parsing the `install_rust.sh` script. This has been
changed to use bash, which should fix the problem.

* Remove install script

The install script introduces a lot of additional overhead during the
build process. So instead of one docker image with an install script
there now are two images, one for stable and one for nightly

* Correct cargo-fmt path

* Set cargo bin path in docker

* Add travis badge to README.md

* Change travis badge location

* Revert "Add basic Time component"

This reverts commit 6b516ff1fd6f0a283f645f5b3caf4c5cb20e24aa.